+++ PersonSearchDetails/TestSettings.cs
@@ -20,7 +20,7 @@ namespace WCCIS.Specs
             FireEvents = false,
             Headless = true,
             UserAgent = false,
-            //DriversPath = @"""C:\Users\sa286848\Documents\RedCortex\EasyRepro\Microsoft.Dynamics365.UIAutomation.Sample\chromedriver\chromedriver.exe"""
+            // DriversPath = @"""C:\Users\sa286848\Documents\RedCortex\EasyRepro\Microsoft.Dynamics365.UIAutomation.Sample\chromedriver\chromedriver.exe"""
         };
     }
 }
